Adding a conftest file for shared fixtures. This file also sets up the path for watchtower.
